+++ packages/@react-elf/ui/src/index.ts
@@ -13,3 +13,4 @@ export * from "@react-elf/dialog";
 export * from "@react-elf/toolbar";
 export * from "@react-elf/tools";
 export * from "@react-elf/ghost";
+export * from "@react-elf/badge";